Refactor Link component
